requirements.txt
fastapi
pydantic
psycopg2
redis
python-dotenv
uvicorn
asyncio
passlib
PyJWT